feat: baseline audit and server gameplay API boundary

+++ engine/src/sync/public/snapshot.h
@@ -24,8 +24,8 @@ struct EntitySnapshot {
     glm::vec3 velocity{0.0f, 0.0f, 0.0f};
     glm::quat rotation{1.0f, 0.0f, 0.0f, 0.0f};
     std::uint16_t hp = 0;
-    std::uint8_t state = 0;
-    std::uint8_t flags = 0;
+    std::uint16_t state = 0;
+    std::uint32_t flags = 0;
 };
 
 struct WorldSnapshot {
